fix: resolve lint errors and warnings
