+++ src/shared/lib/baseQuery.ts
@@ -27,8 +27,6 @@ export const baseQueryWithReauth: BaseQueryFn<
   FetchBaseQueryError
 > = async (args, api, extraOptions) => {
   await mutex.waitForUnlock();
-
-
   let result = await baseQuery(args, api, extraOptions);
   const token = localStorage.getItem('accessToken');
   if (result.error && result.error.status === 401 && token) {
@@ -60,8 +58,6 @@ export const baseQueryWithReauth: BaseQueryFn<
       await mutex.waitForUnlock();
       result = await baseQuery(args, api, extraOptions);
     }
-
   }
-
   return result;
 };
