+++ src/widgets/side-bar/SidebarData.tsx
@@ -1,5 +1,3 @@
-import { useGetMeQuery } from '@/features/auth/api/authApi';
-import { IconSprite } from '@/shared/ui';
 import { SpriteName } from 'public/icons/spriteNames';
 
 type SidebarItem = {
